requirements.txt
fastapi
uvicorn
pydantic
crewai
langchain-openai
playwright
pymongo
google-search-results
python-dotenv
faiss-cpu

